extracts renderSubMenuItems function into module
