+++ components/src/NavBar/MenuTrigger.js
@@ -3,9 +3,8 @@ import styled from "styled-components";
 import PropTypes from "prop-types";
 import theme from "../theme";
 import { Icon } from "../Icon";
-import SubMenuTrigger from "./SubMenuTrigger";
-import SubMenuLink from "./SubMenuLink";
 import { DropdownMenu } from "../DropdownMenu";
+import renderSubMenuItems from "./renderSubMenuItems";
 
 const StyledButton = styled.button({
   display: "block",
@@ -41,66 +40,6 @@ const MenuTriggerButton = React.forwardRef(({ name, ...props }, ref) => (
 MenuTriggerButton.propTypes = {
   name: PropTypes.string.isRequired
 };
-
-const ApplySubMenuLinkStyles = styled.li({
-  color: theme.colors.black,
-  whiteSpace: "nowrap",
-  borderColor: "transparent",
-  backgroundColor: "transparent",
-  textDecoration: "none",
-  verticalAlign: "middle",
-  lineHeight: theme.lineHeights.base,
-  transition: ".2s",
-  fontSize: theme.fontSizes.medium,
-  "> *": {
-    display: "block",
-    color: theme.colors.darkBlue,
-    textDecoration: "none",
-    padding: `${theme.space.x1} ${theme.space.x2}`,
-    "&:hover, &:focus": {
-      outline: "none",
-      backgroundColor: theme.colors.lightGrey
-    },
-    "&:disabled": {
-      opacity: ".5"
-    }
-  }
-});
-
-const renderSubMenuTrigger = (subMenuItem, linkOnClick) => (
-  <li style={{ whiteSpace: "nowrap" }} key={subMenuItem.name}>
-    <SubMenuTrigger linkOnClick={linkOnClick} name={subMenuItem.name} menuData={subMenuItem.items} />
-  </li>
-);
-
-const renderSubMenuLink = (subMenuItem, linkOnClick) => (
-  <li style={{ whiteSpace: "nowrap" }} key={subMenuItem.name}>
-    <SubMenuLink onClick={linkOnClick} href={subMenuItem.href}>
-      {subMenuItem.name}
-    </SubMenuLink>
-  </li>
-);
-
-const renderCustom = (subMenuItem, linkOnClick) => (
-  <ApplySubMenuLinkStyles key={subMenuItem.name} onClick={linkOnClick}>
-    {subMenuItem.render()}
-  </ApplySubMenuLinkStyles>
-);
-
-const getRenderFunction = subMenuItem => {
-  if (subMenuItem.items) {
-    return renderSubMenuTrigger;
-  } else if (subMenuItem.href) {
-    return renderSubMenuLink;
-  } else if (subMenuItem.render) {
-    return renderCustom;
-  } else {
-    return () => null;
-  }
-};
-
-const renderSubMenuItems = (subMenuItems, linkOnClick) =>
-  subMenuItems.map(subMenuItem => getRenderFunction(subMenuItem)(subMenuItem, linkOnClick));
 
 const MenuTrigger = props => {
   const { menuData, name, ...otherProps } = props;
